feat: Add initial structure
The main goal of this structure is that there is not much template files but rather a start project that can be worked on an re-used for new projects. The old project should also be able to benefit from the tempalte and ofr that I hope that Copier would be a good fit

There is a design decision to put everything in the root folder as I feel that putting it into "project_dir" just overcomlicates things.