feat(frontend): Return of the 2 containers setup
So previously we removed this setup as it was simplier to maintain with just one Django container and no frontend build step. However as we also want to use TailwindCSS most of the time, and the pure Python implementation is quite slow and unreliable, decided to return to the initial idea of very basic Frontend that is used only to generate styles/emails on the build step.