feat(users): Port back Django account model
We are using the custom user model, however at some point we decided to get rid of the `first_name` and `last_name` fields. Now this does not work well with Wagtail setup - the user profile editing form does not work. Instead of customizing the admin forms as well, we decided that it is simplier to just add these two fields back and have a good working system.