feat(users): Add Django login/logout/reset password views
When working on MVT it is greatly appreciated if something is already done on the framework side. In our case it is great that Django has a built-in authentication framework, and it even has built-in but ugly as heck templates.
Let's reuse this and just style templates instead of reinventing the weel.